move callback logic to hook
