+++ store/useAuthManagerStore.ts
@@ -1,14 +1,14 @@
 import OAuthManager from "@/services/OAuthManager";
 import { create } from "zustand";
 
-type AuthManager = {
+type AuthManagerStore = {
   manager: OAuthManager | null;
   setManager: (manager: OAuthManager) => void;
   hasAccessToken: boolean;
   setHasAccessToken: (hasAccessToken: boolean) => void;
 };
 
-export const useAuthManager = create<AuthManager>((set) => ({
+export const useAuthManagerStore = create<AuthManagerStore>((set) => ({
   manager: null,
   setManager: (manager: OAuthManager) => set({ manager }),
   hasAccessToken: false,
